ONB-129 feat: 플레이 횟수 타입 Long으로 변경

+++ data/src/main/java/com/yapp/bol/data/model/user/TotalMatchCountResponse.kt
@@ -1,5 +1,5 @@
 package com.yapp.bol.data.model.user
 
 data class TotalMatchCountResponse(
-    val matchCount: Int,
+    val matchCount: Long,
 )
