+++ domain/src/main/java/com/yapp/bol/domain/model/user/TotalMatchCountItem.kt
@@ -1,3 +1,5 @@
 package com.yapp.bol.domain.model.user
 
-data class TotalMatchCountItem(val matchCount: Int)
+data class TotalMatchCountItem(
+    val matchCount: Long,
+)
